Remove public declarations
This package is only intended to be accessed through `rmw_implementation`. As such none of the headers in this package were intended to be part of the public interface. The link interface of this package is that provided by `rmw_implementation`.
