+++ rmw_cyclonedds_cpp/src/macros.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RMW_CYCLONEDDS_CPP__MACROS_HPP_
-#define RMW_CYCLONEDDS_CPP__MACROS_HPP_
+#ifndef MACROS_HPP_
+#define MACROS_HPP_
 
 #include <limits>
 #include <string>
@@ -33,4 +33,4 @@
     } \
   };
 
-#endif  // RMW_CYCLONEDDS_CPP__MACROS_HPP_
+#endif  // MACROS_HPP_
