+++ rmw_cyclonedds_cpp/src/rmw_version_test.hpp
@@ -11,8 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#ifndef RMW_CYCLONEDDS_CPP__RMW_VERSION_TEST_HPP_
-#define RMW_CYCLONEDDS_CPP__RMW_VERSION_TEST_HPP_
+#ifndef RMW_VERSION_TEST_HPP_
+#define RMW_VERSION_TEST_HPP_
 
 /* True if the version of RMW is at least major.minor.patch */
 #define RMW_VERSION_GTE(major, minor, patch) ( \
@@ -21,4 +21,4 @@
         minor < RMW_VERSION_MINOR || ( \
           minor == RMW_VERSION_MINOR && patch <= RMW_VERSION_PATCH))))
 
-#endif  // RMW_CYCLONEDDS_CPP__RMW_VERSION_TEST_HPP_
+#endif  // RMW_VERSION_TEST_HPP_
